# contracts/openapi/latest/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "properties": {
              "kind": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "type": "object"
          }
        },
        "type": "object"
      },
...
      "LocalizedText": {
        "additionalProperties": {
          "type": "string"
        },
        "type": "object"
      },
...
      "ModelDefinition": {
        "properties": {
          "create_permission_key": {
            "type": "string"
          },
          "default_sort": {
            "type": "string"
          },
          "display_name": {
            "type": "string"
          },
          "display_name_i18n": {
            "$ref": "#/components/schemas/LocalizedText"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/ModelFieldDefinition"
            },
            "type": "array"
          },
          "key": {
            "type": "string"
          },
          "list_permission_key": {
            "type": "string"
          },
          "owner_module_key": {
            "type": "string"
          },
          "read_permission_key": {
            "type": "string"
          },
          "relations": {
            "items": {
              "$ref": "#/components/schemas/ModelRelationDefinition"
            },
            "type": "array"
          },
          "update_permission_key": {
            "type": "string"
          },
          "version": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ModelListResponse": {
        "properties": {
          "definition": {
            "$ref": "#/components/schemas/ModelDefinition"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ModelRecord"
            },
            "type": "array"
          },
          "related_definitions": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ModelDefinition"
            },
            "type": "object"
          },
          "total": {
            "type": "integer"
          }
        },
        "type": "object"
      },
      "ModelMutationRequest": {
        "properties": {
          "relations": {
            "additionalProperties": true,
            "type": "object"
          },
          "values": {
            "additionalProperties": true,
            "type": "object"
          }
        },
        "type": "object"
      },
      "ModelMutationResponse": {
        "properties": {
          "record": {
            "$ref": "#/components/schemas/ModelRecord"
          },
          "related": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/ModelRecord"
              },
              "type": "array"
            },
            "type": "object"
          }
        },
        "type": "object"
      },
      "ModelRecord": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "created_by": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "model_key": {
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_by": {
            "type": "string"
          },
          "values": {
            "additionalProperties": true,
            "type": "object"
          },
          "version": {
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "UIModelDataResponse": {
        "additionalProperties": true,
        "properties": {
          "items": {
            "items": {
              "additionalProperties": true,
              "description": "UI model item",
              "type": "object"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "OpenAPI document for the runtime generic platform APIs. Registered modules become available through the generic model, document, UI contract, search, and offline endpoints. Custom module-specific controller namespaces are not auto-generated.",
    "title": "Orbyte Platform API",
    "version": "0.1.0-dev"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/models/{modelKey}": {
      "get": {
        "description": "Lists records for the given manifest-registered model key.",
        "parameters": [
          {
            "description": "Registered model key.",
            "in": "path",
            "name": "modelKey",
            "required": true,
            "schema": {
              "enum": [
                "absence_code",
                "accounting_period",
                "accounting_period_task",
                "approval_delegation_rule",
                "approval_policy",
                "approval_policy_stage",
                "approver_group",
                "approver_group_member",
                "asset_disposal",
                "asset_impairment",
                "asset_revaluation",
                "asset_transfer",
                "attendance_adjustment",
                "attendance_day",
                "attendance_event",
                "bank_import_preset",
                "bank_import_template",
                "bank_match_rule",
                "bank_reconciliation",
                "bank_reconciliation_match",
                "bank_statement",
                "bank_statement_import_run",
                "bank_statement_line",
                "collection_case",
                "commercial_account",
                "commercial_item",
                "commercial_item_category",
                "commercial_price_list",
                "commercial_price_list_item",
                "commercial_product",
                "commercial_tax_code",
                "commercial_tax_profile",
                "commercial_uom",
                "commercial_variant_dimension",
                "commercial_variant_value",
                "cost_center",
                "crm_activity",
                "crm_assignment_rule",
                "crm_lead",
                "crm_opportunity",
                "crm_queue",
                "crm_sla_policy",
                "crm_ticket",
                "crm_ticket_activity",
                "crm_ticket_comment",
                "customer_profile",
                "department",
                "discount_rule",
                "employee_assignment",
                "employee_compensation_profile",
                "employee_leave_profile",
                "employee_payroll_profile",
                "employee_profile",
                "employee_role_eligibility",
                "employee_spend_profile",
                "expense_category",
                "expense_policy",
                "expense_rate_rule",
                "finance_account",
                "fixed_asset",
                "fixed_asset_schedule",
                "gift_card",
                "gift_card_transaction",
                "inventory_batch",
                "inventory_cost_layer",
                "inventory_count_session",
                "inventory_reconciliation_case",
                "inventory_valuation_snapshot",
                "journal_run",
                "journal_template",
                "leave_accrual_run",
                "leave_balance_account",
                "leave_balance_adjustment",
                "leave_balance_entry",
                "leave_entitlement_rule",
                "leave_policy",
                "leave_request",
                "organization_unit",
                "overtime_request",
                "party",
                "party_address",
                "party_contact",
                "party_statement_run",
                "pay_component",
                "payment_method",
                "payroll_contribution_rule",
                "payroll_period",
                "payroll_remittance_profile",
                "payroll_tax_rule",
                "planning_proposal",
                "planning_run",
                "pos_register",
                "pos_sale",
                "pos_shift",
                "pos_store",
                "pos_tender_reconciliation",
                "pos_tender_settlement",
                "pos_tender_type",
                "prepaid_expense",
                "prepaid_schedule",
                "production_bom",
                "production_bom_version",
                "production_cost_capture",
                "production_cost_rate",
                "production_output_allocation",
                "production_routing",
                "production_routing_step",
                "production_variance_case",
                "production_work_center",
                "promotion_campaign",
                "promotion_code",
                "promotion_redemption",
                "remittance_authority",
                "remittance_obligation_type",
                "remittance_schedule_rule",
                "salary_structure",
                "salary_structure_line",
                "settlement_exception",
                "shift_template",
                "store_credit_account",
                "store_credit_transaction",
                "travel_policy",
                "treasury_account",
                "treasury_exception",
                "treasury_transfer",
                "vendor_item_profile",
                "vendor_profile",
                "vendor_statement_run",
                "warehouse",
                "work_calendar",
                "workforce_roster",
                "workforce_roster_slot"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelListResponse"
                }
              }
            },
            "description": "Model list"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          }
        },
        "summary": "List model records",
        "tags": [
          "Models"
        ],
        "x-orbyte-support-level": "public-headless"
      },
      "post": {
        "description": "Creates a record through the generic model API for the given model key.",
        "parameters": [
          {
            "description": "Registered model key.",
            "in": "path",
            "name": "modelKey",
            "required": true,
            "schema": {
              "enum": [
                "absence_code",
                "accounting_period",
                "accounting_period_task",
                "approval_delegation_rule",
                "approval_policy",
                "approval_policy_stage",
                "approver_group",
                "approver_group_member",
                "asset_disposal",
                "asset_impairment",
                "asset_revaluation",
                "asset_transfer",
                "attendance_adjustment",
                "attendance_day",
                "attendance_event",
                "bank_import_preset",
                "bank_import_template",
                "bank_match_rule",
                "bank_reconciliation",
                "bank_reconciliation_match",
                "bank_statement",
                "bank_statement_import_run",
                "bank_statement_line",
                "collection_case",
                "commercial_account",
                "commercial_item",
                "commercial_item_category",
                "commercial_price_list",
                "commercial_price_list_item",
                "commercial_product",
                "commercial_tax_code",
                "commercial_tax_profile",
                "commercial_uom",
                "commercial_variant_dimension",
                "commercial_variant_value",
                "cost_center",
                "crm_activity",
                "crm_assignment_rule",
                "crm_lead",
                "crm_opportunity",
                "crm_queue",
                "crm_sla_policy",
                "crm_ticket",
                "crm_ticket_activity",
                "crm_ticket_comment",
                "customer_profile",
                "department",
                "discount_rule",
                "employee_assignment",
                "employee_compensation_profile",
                "employee_leave_profile",
                "employee_payroll_profile",
                "employee_profile",
                "employee_role_eligibility",
                "employee_spend_profile",
                "expense_category",
                "expense_policy",
                "expense_rate_rule",
                "finance_account",
                "fixed_asset",
                "fixed_asset_schedule",
                "gift_card",
                "gift_card_transaction",
                "inventory_batch",
                "inventory_cost_layer",
                "inventory_count_session",
                "inventory_reconciliation_case",
                "inventory_valuation_snapshot",
                "journal_run",
                "journal_template",
                "leave_accrual_run",
                "leave_balance_account",
                "leave_balance_adjustment",
                "leave_balance_entry",
                "leave_entitlement_rule",
                "leave_policy",
                "leave_request",
                "organization_unit",
                "overtime_request",
                "party",
                "party_address",
                "party_contact",
                "party_statement_run",
                "pay_component",
                "payment_method",
                "payroll_contribution_rule",
                "payroll_period",
                "payroll_remittance_profile",
                "payroll_tax_rule",
                "planning_proposal",
                "planning_run",
                "pos_register",
                "pos_sale",
                "pos_shift",
                "pos_store",
                "pos_tender_reconciliation",
                "pos_tender_settlement",
                "pos_tender_type",
                "prepaid_expense",
                "prepaid_schedule",
                "production_bom",
                "production_bom_version",
                "production_cost_capture",
                "production_cost_rate",
                "production_output_allocation",
                "production_routing",
                "production_routing_step",
                "production_variance_case",
                "production_work_center",
                "promotion_campaign",
                "promotion_code",
                "promotion_redemption",
                "remittance_authority",
                "remittance_obligation_type",
                "remittance_schedule_rule",
                "salary_structure",
                "salary_structure_line",
                "settlement_exception",
                "shift_template",
                "store_credit_account",
                "store_credit_transaction",
                "travel_policy",
                "treasury_account",
                "treasury_exception",
                "treasury_transfer",
                "vendor_item_profile",
                "vendor_profile",
                "vendor_statement_run",
                "warehouse",
                "work_calendar",
                "workforce_roster",
                "workforce_roster_slot"
              ],
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelMutationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelMutationResponse"
                }
              }
            },
            "description": "Model created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          }
        },
        "summary": "Create model record",
        "tags": [
          "Models"
        ],
        "x-orbyte-support-level": "public-headless"
      }
    },
...
    "/ui/data/models": {
      "get": {
        "description": "Returns model data for the generic user workspace.",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UIModelDataResponse"
                }
              }
            },
            "description": "UI model data"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          }
        },
        "summary": "Get model list for UI",
        "tags": [
          "UI Contracts"
        ],
        "x-orbyte-support-level": "public-headless"
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
}